+++ packages/react-native-eyecandy-icons/src/icons/Share.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import { Path } from 'react-native-svg';
+
+import { SvgIconContainer } from '../core';
+import { SvgIconProps } from '../types';
+
+export const Share = (props: SvgIconProps) => (
+  <SvgIconContainer {...props}>
+    <Path
+      d="M6 15a3 3 0 1 0 0-6 3 3 0 0 0 0 6ZM18 9a3 3 0 1 0 0-6 3 3 0 0 0 0 6ZM18 21a3 3 0 1 0 0-6 3 3 0 0 0 0 6ZM8.7 10.7l6.6-3.4M8.7 13.3l6.6 3.4"
+      strokeWidth={1.5}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </SvgIconContainer>
+);