+++ packages/react-native-eyecandy/src/components/TextInput/TextInput.tsx
@@ -4,16 +4,19 @@ import {
   TextInput as RNTextInput,
   StyleProp,
   TextStyle,
+  StyleSheet,
+  ViewStyle,
+  TouchableWithoutFeedback,
 } from 'react-native';
 
 import { EyeCheck, EyeOff } from '@nomada-sh/react-native-eyecandy-icons';
 
 import { Body } from '../../typography';
+import { extractOnlyTextStyles } from '../../utils/extractOnlyTextStyles';
 
 import IconTouchable from './IconTouchable';
 import styles from './styles';
 import { TextInputStyles, TextInputProps, TextInputHandle } from './types';
-import useSecureTextEntry from './useSecureTextEntry';
 import useStyles from './useStyles';
 
 const DEFAULT_CUSTOM_STYLES: TextInputStyles = {};
@@ -38,11 +41,14 @@ export const TextInput = React.forwardRef<TextInputHandle, TextInputProps>(
       onFocus,
       onBlur,
       autoFocus,
-      hideIconLeftUnfocused,
-      hideIconRightUnfocused,
-      renderValueAsTextUnfocused,
+      hideIconLeftWhenUnfocused,
+      hideIconRightWhenUnfocused,
+      renderInputAsTextWhenUnfocused,
       value,
       editable = true,
+      secureTextEntry,
+      onSecureTextEntryChange,
+      numberOfLines,
       ...inputProps
     } = props;
 
@@ -76,10 +82,10 @@ export const TextInput = React.forwardRef<TextInputHandle, TextInputProps>(
     let iconLeftVisible = !!iconLeft;
     let iconRightVisible = !!iconRight;
 
-    if (iconLeftVisible && hideIconLeftUnfocused && !focused)
+    if (iconLeftVisible && hideIconLeftWhenUnfocused && !focused)
       iconLeftVisible = false;
 
-    if (iconRightVisible && hideIconRightUnfocused && !focused)
+    if (iconRightVisible && hideIconRightWhenUnfocused && !focused)
       iconRightVisible = false;
 
     const dynamicStyles = useStyles({
@@ -116,9 +122,6 @@ export const TextInput = React.forwardRef<TextInputHandle, TextInputProps>(
       )
     ) : null;
 
-    const { secureTextEntry, onPressSecureTextEntryToggle } =
-      useSecureTextEntry(props);
-
     const handlePressIconLeft = () => {
       if (onPressIconLeft) onPressIconLeft();
       if (focusOnLeftIconPress) focus();
@@ -141,6 +144,30 @@ export const TextInput = React.forwardRef<TextInputHandle, TextInputProps>(
         : customStyles.input,
       inputStyle,
     ];
+
+    const flattenInputStyles = renderInputAsTextWhenUnfocused
+      ? StyleSheet.flatten(inputStyles)
+      : undefined;
+
+    const inputTextStyles = flattenInputStyles
+      ? extractOnlyTextStyles(flattenInputStyles)
+      : undefined;
+
+    let justifyContent: ViewStyle['justifyContent'] = undefined;
+
+    if (inputTextStyles) {
+      switch (inputTextStyles.textAlignVertical) {
+        case 'bottom':
+          justifyContent = 'flex-end';
+          break;
+        case 'top':
+          justifyContent = 'flex-start';
+          break;
+        default:
+          justifyContent = 'center';
+          break;
+      }
+    }
 
     const placeholder =
       placeholderProp && required ? `${placeholderProp} *` : placeholderProp;
@@ -168,24 +195,36 @@ export const TextInput = React.forwardRef<TextInputHandle, TextInputProps>(
           ]}
           icon={iconLeft}
           color={dynamicStyles.icon.color}
-          hideUnfocused={hideIconLeftUnfocused}
+          hideUnfocused={hideIconLeftWhenUnfocused}
         />
 
         {/* Left Component */}
         {inputLeft}
 
-        {renderValueAsTextUnfocused && !focused ? (
-          <Body
-            style={[
-              {
-                textAlignVertical: 'center',
-              },
-              inputStyles,
-            ]}
-            onPress={focus}
-          >
-            {value ? value : placeholder}
-          </Body>
+        {/* Workaround for android can't scroll when textAlign is 'center' or 'right': https://github.com/facebook/react-native/issues/25594 */}
+        {renderInputAsTextWhenUnfocused && !focused ? (
+          <TouchableWithoutFeedback onPress={focus}>
+            <View style={[{ justifyContent }, inputStyles]}>
+              <Body
+                style={[
+                  inputTextStyles,
+                  {
+                    color: value
+                      ? dynamicStyles.input.color
+                      : dynamicStyles.placeholder.color,
+                    paddingBottom: 1.5,
+                  },
+                ]}
+                numberOfLines={numberOfLines}
+              >
+                {value
+                  ? secureTextEntry
+                    ? new Array(value.length).fill('*').join('')
+                    : value
+                  : placeholder}
+              </Body>
+            </View>
+          </TouchableWithoutFeedback>
         ) : null}
 
         <RNTextInput
@@ -195,11 +234,12 @@ export const TextInput = React.forwardRef<TextInputHandle, TextInputProps>(
           placeholderTextColor={dynamicStyles.placeholder.color}
           style={[
             inputStyles,
-            renderValueAsTextUnfocused && !focused
+            renderInputAsTextWhenUnfocused && !focused
               ? {
                   position: 'absolute',
                   bottom: 0,
                   width: 0,
+                  height: 0,
                 }
               : undefined,
           ]}
@@ -216,6 +256,7 @@ export const TextInput = React.forwardRef<TextInputHandle, TextInputProps>(
           }}
           secureTextEntry={secureTextEntry}
           ref={inputRef}
+          numberOfLines={numberOfLines}
           {...inputProps}
         />
 
@@ -234,14 +275,17 @@ export const TextInput = React.forwardRef<TextInputHandle, TextInputProps>(
           ]}
           icon={iconRight}
           color={dynamicStyles.icon.color}
-          hideUnfocused={hideIconRightUnfocused}
+          hideUnfocused={hideIconRightWhenUnfocused}
         />
 
         {/* Secure Text Entry Toggle */}
         {showSecureTextEntryToggle ? (
           <IconTouchable
             focused={focused}
-            onPress={onPressSecureTextEntryToggle}
+            onPress={() =>
+              onSecureTextEntryChange &&
+              onSecureTextEntryChange(!secureTextEntry)
+            }
             style={[
               styles.rightIconContainer,
               customStyles.rightIconContainer instanceof Function