+++ packages/react-native-eyecandy/src/components/TextInput/index.ts
@@ -2,6 +2,5 @@ export * from './TextInput';
 export * from './types';
 
 export { default as IconTouchable } from './IconTouchable';
-export { default as useSecureTextEntry } from './useSecureTextEntry';
 export { default as useStyles } from './useStyles';
 export { default as styles } from './styles';