+++ packages/react-native-eyecandy/src/components/TextInputEmail/index.ts
@@ -1,1 +1,1 @@
-export { default } from './TextInputEmail';
+export * from './TextInputEmail';