+++ packages/react-native-eyecandy/src/components/TextInputPassword/index.ts
@@ -1,1 +1,1 @@
-export { default } from './TextInputPassword';
+export * from './TextInputPassword';